Preliminary declaration parser
Major changes include the addition of the ParseDec module, which for
now is capable of parsing functions and struct declarations.  Classes,
interfaces, protocols, and anything involving templates is forthcoming.

Minor changes include the addition of the “arrow call” syntax for
chaining function calls where the first argument of a function call is
another function call, and the introduction of the “do” keyword for
trailing statement blocks which terminate when their enclosing block
terminates.