Consolidated "parseCompoundStat"
parseCompoundStat is no longer a redundant and independent statement
parser from parseStat, but is instead a group of sub-choices for
parseStat.  This has two benefits: first of all, statements cannot be
added to parseComoundStat and accidentally left out of parseStat (which
should never happen); additionally, this decreases the probability of
accidentally neglecting to add a compound statement to
parseCompoundStat, as happened in the previous commit for extract
statements.