Close #11 — Shorthand for common types
The “T[]” shorthand has been added for lists, and the “V[K]” syntax has
been added for dictionaries.  The “V[K]” syntax involves the reversal
of the key-value order relative to how it appears in the desugared
template instantiation, and is therefore stylistically questionable.
This syntax may be subject to change in the future.