Close #7 — Parse "given" clauses
The actual syntax for “given” clauses ended up being different from the
planned syntax in the github issue.  Protocol and function requirements
are no longer syntactically indistinguishable, and the entire given
clause has been changed to be much more consistent with the body of a
“protocol” declaration.  In particular, “given” clauses now use curly
braces, and require that every requirement be terminated with a
semicolon.