Improved main function to allow parsing whole files
The current main function is completely temporary and exists only for
the purpose of debugging the parser, but it is nonetheless nice to be
able to test whole files, rather than only being able to work with
single lines.

Files to parse can be specified as a command-line argument to the
executable.  When a file is not specified, the executable defaults to
the old behavior of reading single lines as modules.  The file-to-parse
argument cannot be specified when the executable is invoked through
“./run.sh”