Close #17 — Parse named infix operators
Named infix operators are supported both in expressions and in
update-assignment statements (like, for example, “v1 `cross`= v2”).  In
expressions containing multiple binary operators, named infix operators
have the lowest precedence (even lower than `&&`, but not lower than
`=`, because `=` is not an operator, unlike in C).  Named infix
operators are left-associative.  In the future, update assignment with
named infix operators may support taking multiple argument on the
right-hand-side (unlike other update assignment operators), but this is
currently not supported because it would introduce a special case into
the update assignment parser.