Close #24 — Parse index expressions
“foo[bar]” is now exact syntactic sugar for “foo.index(bar)”.  This
means that index expressions are implicitly subject to namespace
resolution rules looking for an accessor called “index”.  A similar
situation occurs in other places in the parser, such as with arithmetic
operators and list/dictionary type shorthand.  In all cases, the
implicit involvement of namespace resolution in a piece of syntax which
seems to contain no names is stylistically questionable, and may be
subject to change at a later date.